check-branches: Automatically fill in Config.host_owner
The correct owner can be extracted from the remote origin's URL:

    git@github.com:<owner>/<repo_name>

It's not useful to require it to be configured, since it's readily
available. Plus, not configuring it breaks check-branches which cannot
build proper URLs using the default `host-url` template (`{owner}` is
replaced by `None`).

Remove any possibilities to configure host-owner.

The parsing is slightly more complex to accomodate both HTTP and SSH
forms, the most common when using GitHub.