req: Remove optional dependency on `toml`
It was not necessary with Python >= 3.11, which ships with the
`tomllib` package in the standard library.

Also, the extra requirement could be a bit confusing, especially for
non-Python developers. "Do I need it now, will I need it later?"

For the sake of simplicity, the extra requirement has been removed.
We now always install `toml` on Python versions earlier than 3.11 (and
never on 3.11 and above).